pyproject.toml
[project]
name = "vla-distributed-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=15.1.0",
    "huggingface-hub[cli,hf-transfer]==0.35.3",
    "ipywidgets>=8.1.8",
    "lerobot==0.4.3",
    "python-dotenv>=1.2.2",
    "ray[data,train]>=2.54.1",
    "sentencepiece>=0.2.1",
    "transformers",
    "wandb>=0.24.2",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git", rev = "fix/lerobot_openpi" }
